+++ packages/core/src/modal-suggest/ModalSuggestOptionModel.ts
@@ -1,5 +1,0 @@
-import {OptionModel} from '../option'
-
-export default interface ModalSuggestOptionModel<V> extends OptionModel<V> {
-  suggest?: string
-}